mlflow/inference_engine/requirements.txt
mlflow>=2.10.0
torch>=2.0.0,<2.11.0
scikit-learn>=1.3.0
pandas>=2.0.0
numpy>=2.4.0,<3.0.0
psycopg2-binary>=2.9.0
sqlalchemy>=2.0.0
kafka-python>=2.0.2